+++ app/controllers/Home.php
@@ -1,0 +1,9 @@
+<?php
+
+class Home extends Controller
+{
+    public function index(): void
+    {
+        echo 'this is the home comtroller';
+    }
+}
