+++ app/core/config.php
@@ -1,0 +1,8 @@
+<?php
+
+function show($stuff)
+{
+    echo '<pre>';
+    print_r($stuff);
+    echo '</pre>';
+}
