+++ app/core/function.php
@@ -1,0 +1,1 @@
+
