+++ app/core/init.php
@@ -1,0 +1,8 @@
+<?php
+
+require 'config.php';
+require 'function.php';
+require 'Database.php';
+require 'Model.php';
+require 'Controller.php';
+require 'App.php';
