+++ app/views/404.view.php
@@ -1,0 +1,1 @@
+Page not found!
